Separate Hawk code
This code is large enough, so it also deserves its own file
