+++ dma.h
@@ -1,0 +1,1 @@
+void hawk_set_dma(unsigned mode);
